Fix: Properly handle large chunks and track payload size in chunk headers
- Resolved issues with sending large chunks from Client to Server by ensuring proper handling of chunk boundaries and payload sizes.
- Updated the logic to track the payload size in the chunk headers on the wire, ensuring accurate data transmission and alignment with protocol expectations.
- Replaced `global_no_signal` with `global_no_chunk_header` to correctly represent the chunk header type in relevant parts of the code.
- Enhanced `ClientStream::append_data` to handle partial chunks more robustly, ensuring data is appended correctly until the chunk is complete.
- Adjusted `write_streams` to account for payload size limits and ensure proper segmentation of data for transmission.
- Updated test cases in `shared_chunk_tester.cpp` to reflect the new chunk header type and validate the handling of payload sizes.
- Improved logging and error handling to provide better insights during debugging and testing.

These changes ensure reliable and efficient handling of large data chunks, improving the robustness of the Client-Server communication.

+++ common/include/tcp_communication.h
@@ -38,7 +38,7 @@ private:
         id.datalen = 2;
         id.data[0] = 0;
         id.data[1] = 11;
-        return StreamIdentifier(id, 42ul);
+        return StreamIdentifier(id, static_cast<uint16_t>(42));
     }
 
     void terminate() {
